@@ -9,6 +9,7 @@ using System.Text;
 using System;
 using Microsoft.Extensions.Configuration;
 using System.IdentityModel.Tokens.Jwt;
+using AutoMapper;
 
 namespace DatingApp.API.Controllers
 {
@@ -18,8 +19,10 @@ namespace DatingApp.API.Controllers
   {
     private readonly IAuthRepository _repo;
     private readonly IConfiguration _config;
-    public AuthController(IAuthRepository repo, IConfiguration config)
+    private readonly IMapper _mapper;
+    public AuthController(IAuthRepository repo, IConfiguration config, IMapper mapper)
     {
+      _mapper = mapper;
       _config = config;
       _repo = repo;
 
@@ -63,7 +66,8 @@ namespace DatingApp.API.Controllers
 
       var creds = new SigningCredentials(key, SecurityAlgorithms.HmacSha512Signature);
 
-      var tokenDescriptor = new SecurityTokenDescriptor {
+      var tokenDescriptor = new SecurityTokenDescriptor
+      {
         Subject = new ClaimsIdentity(claims),
         Expires = DateTime.Now.AddDays(1),
         SigningCredentials = creds
@@ -73,12 +77,16 @@ namespace DatingApp.API.Controllers
 
       var token = tokenHandler.CreateToken(tokenDescriptor);
 
-      return Ok(new {
-          token = tokenHandler.WriteToken(token)
+      var user = _mapper.Map<UserForListDTO>(userFromRepo);
+
+      return Ok(new
+      {
+        token = tokenHandler.WriteToken(token),
+        user
       });
 
     }
 
-    
+
   }
 }
