@@ -3,21 +3,29 @@ using Microsoft.AspNetCore.Http;
 
 namespace DatingApp.API.Helpers
 {
-    public static class Extensions
+  public static class Extensions
+  {
+    public static void AddApplicationError(this HttpResponse response, string message)
     {
-        public static void AddApplicationError(this HttpResponse response, string message) {
-            response.Headers.Add("Application-Error", message);
-            response.Headers.Add("Access-Control-Expose-Headers", "Application-Error");
-            response.Headers.Add("Access-Control-Allow-Origin", "*");
-        }
-
-        public static int CalculateAge(this DateTime theDateTime) {
-            var age = DateTime.Today.Year - theDateTime.Year;
-            if (theDateTime.AddYears(age) > DateTime.Today) {
-                age--;
-            }
-
-            return age;
-        }
+      response.Headers.Add("Application-Error", message);
+      response.Headers.Add("Access-Control-Expose-Headers", "Application-Error");
+      response.Headers.Add("Access-Control-Allow-Origin", "*");
     }
+
+    public static int CalculateAge(this DateTime theDateTime)
+    {
+      var age = DateTime.Today.Year - theDateTime.Year;
+      if (theDateTime.AddYears(age) > DateTime.Today)
+      {
+        age--;
+      }
+
+      return age;
+    }
+
+    public static string SayHello(string name)
+    {
+      return ("Hello " + name + "!");
+    }
+  }
 }
